backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = {extras = ["standard"], version = "^0.17.0"}
boto3 = "^1.18.0"
python-dotenv = "^1.0.0"
databases = "^0.5.5"
aiosqlite = "^0.17.0"
aws-cdk-lib = "==2.166.0"
greenlet = ">=3.0.0,<3.3.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
flake8 = "^3.9.2"
